Adding user-friendly controls for max time.
Adds 3 QSpinBox's next to max time for days, hours, minutes. Also a QDoubleSpinBox for anybody wanting to use fractions of a minute. The number of decimals for that box is tied to the size of diffusion_dt (decrease diffusion_dt and you can get more decimals).

The value of Max Time (which is all that matters at save time) is synced with these so that they mutually update one another. Recursive loops are avoided by adding a flag for updating the spin boxes.

This should make the experience of setting max time a bit more user-friendly since we are not accustomed to converting days into minutes.