Fixed bug in loading adhesion affinities.
A similar bug to loading the transformation rates found before. Hence, the same solution. Pass in a reset_mapping flag (the very same as for the interactions tab), and only reset the mappings if that flag is true. I have not attempted to wrap my head fully around the logic of that flag, but it seems to work. Now, both these affinities and the interactions live and die together by this flag.