Rules updates on rename.
First, fixed a bug where renaming a substrate could cause an errant name in the dropdowns in the rules tab. Specifically, if one substrate is a substring of another, e.g. IFN and IFNg or substrate and anti-substrate, renaming the substring version would ALSO rename the superstring version only in the dropdowns. So, IFN-->interferon would cause IFNg to disappear from the dropdowns and be replaced by interferong.

Second, any cell or substrate rename is now automatically reflected in the cell rules table already loaded in studio. It cannot save you if you import one *after* a rename. The functions carefully check that no superstrings are being replaced in the process to avoid a similar issue to that described above. That said, if the user creates substrates with simple/dumb names (simple = "a", "b", "c"; dumb = "with", "dead", "intracellular") issues will arise on rename. See comment in find_and_replace_rule_cell for thoughts on avoiding this. This could be a good reason to NOT do this. Maybe a check for neighboring characters not being whitespace would be a good protection for things like "a".