Handle when names are substrings of formulaic words.
If a cell/substrate name is a substring of a "formulaic word" such as "with", "intracellular", "live", etc. we can now protect against changes to that name causing erroneous changes to the rules table. Specifically, whitespace on either side of a match with the old_name is required to move forward with replacing (or the purported match is at the string end or beginning).

For example, suppose a user names a substrate simply as "a" with a rule for "chemotactic response to a". If the user then renames to "IFN", the string "chemotactic response to a" will be searched for any matches with "a". The first instance, "a" in "chemotactic" will be skipped because the previous character "t" is not a space. The second instance will be identified as the time to replace "a" with "IFN" as there is whitespace before "a" and no characters after "a".