Increment new substrate suffix as necessary.
WHen creating a new substrate, add a suffix to it of the form %02d. If that substrate is already in the list, keep incrementing until you find a name not in the list. Don't go over 99 or else I think it'll just hang as it fails to find a correct name.