Parameter distributions to initialized cells
This gets studio ready for a future PhysiCell commit that will allow parameters to be initialized from distributions. Namely, https://github.com/MathCancer/PhysiCell/pull/203

To do this, I add a Misc tab at the end of the cell def tab. Possible improvement: check that parameter values are sensible/will not raise errors when run in PhysiCell. We're talking min < max and bounds on log-scale parameters are strictly positive (if you want a lognormal par to be arbitrarily small, then don't put on a lower_bound)