show names in plot dropdown rather than ids
no more counting indices to figure out what internalized_total_substrate_37 means! Instead, use the xml to identify them.
same for other mcds variables that are related to substrates.

same is done for cell types, but this (dangerously!) uses the studio to make the substitutions. very possible that users load up a different output folder and then things get weird.
not broken, because we make sure the id exists. but still weird.
but only if the user did a weird thing first (load a different output folder)